feat(security fixes + visual upgrades)

+++ routes/news.js
@@ -16,26 +16,12 @@ router.get("/", async (req, res) => {
         JOIN login ON posts.author_id = login.id 
         ORDER BY created_at DESC;`)
 
-    if (req.session.login && req.session.adminStatus) {
-        res.render("news.njk", {
-            posts: posts,
-            loginbutton: "Log Out",
-            loginURL: "/logout",
-            newpost: "/news/post"
-        })
-    } else if (req.session.login) {
-        res.render("news.njk", {
-            posts: posts,
-            loginbutton: "Log Out",
-            loginURL: "/logout"
-        })
-    } else {
-        res.render("news.njk", {
-            posts: posts,
-            loginbutton: "Log In",
-            loginURL: "/login"
-        })
-    }
+    res.render("news.njk", {
+        posts: posts,
+        login: req.session.login,
+        loginId: req.session.userId,
+        adminStatus: req.session.adminStatus
+    })
 })
 
 router.get("/:id/delete", async (req, res) => {
@@ -50,11 +36,14 @@ router.get("/:id/delete", async (req, res) => {
         return res.status(400).send("Post not found")
     }
 
-    if (req.session.userId == post.author_id) {
+    if (req.session.userId == post.author_id && req.session.adminStatus) {
         await db.run(`DELETE FROM posts WHERE id = ?`, id)
         await db.run("DELETE FROM comments WHERE comments.post_id = ?", id)
+        res.redirect("/news")
+    } else {
+        return res.status(400).send("Something went wrong")
     }
-    res.redirect("/news")
+
 })
 
 router.get("/:id/edit", async (req, res) => {
@@ -64,55 +53,61 @@ router.get("/:id/edit", async (req, res) => {
         return res.status(404).send("Post not found")
     }
 
-    if (req.session.userId == post.author_id) {
-        if (!Number.isInteger(Number(id))) {
-            return res.status(400).send("Invalid Post-ID")
-        }
+    if (!Number.isInteger(Number(id))) {
+        return res.status(400).send("Invalid Post-ID")
+    }
 
+    if (req.session.userId == post.author_id) {
         res.render('edit.njk', {
             message: "Edit Post",
             post: post,
-            loginbutton: "Log Out",
-            loginURL: "/logout",
+            login: req.session.login,
             route: "/news/edit"
         })
     } else {
-        res.redirect("/news")
+        return res.status(400).send("Something went wrong")
     }
 })
 
 router.post('/edit', async (req, res) => {
 
     const { message, id } = req.body
+
     if (!Number.isInteger(Number(id))) {
         return res.status(400).send("Invalid ID")
     }
 
-    await db.run(`UPDATE posts SET message=? where id=?`, message, id)
+    const post = await db.get("SELECT * FROM posts WHERE posts.id = ?", id)
+
+    if (post === undefined) {
+        return res.status(404).send("Post not found")
+    }
+
+    if (req.session.userId == post.author_id) {
+        await db.run(`UPDATE posts SET message=? where id=?`, message, id)
+    }
 
     res.redirect("/news")
 })
 
 router.get("/post", async (req, res) => {
-    const user = await db.all(`SELECT * FROM login;`)
-
     if (req.session.login && req.session.adminStatus == 1) {
         res.render('post.njk', {
             message: "New Post",
-            user: user,
-            loginbutton: "Log Out",
-            loginURL: "/logout",
+            login: req.session.login,
             route: "/news/post",
         })
     } else {
-        res.redirect("/news")
+        return res.status(404).send("Something went wrong")
     }
 })
 
 router.post("/post", async (req, res) => {
     const { message } = req.body
 
-    await db.run(`INSERT INTO posts (author_id, message) values (?, ?)`, req.session.userId, message)
+    if (req.session.login && req.session.adminStatus == 1) {
+        await db.run(`INSERT INTO posts (author_id, message) values (?, ?)`, req.session.userId, message)
+    }
 
     res.redirect("/news")
 })
@@ -124,31 +119,18 @@ router.get("/:id", async (req, res) => {
     }
 
     const post = await db.get("SELECT posts.*, login.name FROM posts JOIN login on posts.author_id = login.id WHERE posts.id = ?", id)
-    console.log(post)
+
     if (post === undefined) {
         return res.status(404).send("Post not found")
     }
     const comments = await db.all("SELECT comments.*, login.name FROM comments JOIN login on comments.author_id = login.id WHERE comments.post_id = ?", id)
-
-    console.log(comments)
-
-    if (req.session.login) {
-        res.render("comments.njk", {
-            message: "",
-            post: post,
-            comments: comments,
-            loginbutton: "Log Out",
-            loginURL: "/logout"
-        })
-    } else {
-        res.render("comments.njk", {
-            message: "",
-            post: post,
-            comments: comments,
-            loginbutton: "Log In",
-            loginURL: "/login"
-        })
-    }
+    res.render("comments.njk", {
+        message: "",
+        post: post,
+        comments: comments,
+        login: req.session.login,
+        loginId: req.session.userId
+    })
 })
 
 router.get("/:id/reply", async (req, res) => {
@@ -167,8 +149,7 @@ router.get("/:id/reply", async (req, res) => {
             message: "Post Comment",
             post: post,
             comments: comments,
-            loginbutton: "Log Out",
-            loginURL: "/logout",
+            login: req.session.login,
             route: "/news/reply",
         })
     } else {
@@ -183,9 +164,12 @@ router.post("/reply", async (req, res) => {
         return res.status(400).send("Invalid Post-ID")
     }
 
-    const result = await db.run("INSERT INTO comments (message, author_id, post_id) values (?, ?, ?)", message, req.session.userId, id)
-    console.log(result)
-    res.redirect(`/news/${id}`)
+    if (req.session.login) {
+        await db.run("INSERT INTO comments (message, author_id, post_id) values (?, ?, ?)", message, req.session.userId, id)
+        res.redirect(`/news/${id}`)
+    } else {
+        return res.status(400).send("Something went wrong")
+    }
 })
 
 router.get("/:id/:commentId/edit", async (req, res) => {
@@ -208,22 +192,37 @@ router.get("/:id/:commentId/edit", async (req, res) => {
             res.render("edit.njk", {
                 message: "Edit Comment",
                 post: comment,
-                loginbutton: "Log Out",
-                loginURL: "/logout",
+                login: req.session.login,
                 route: "/news/comment/edit",
             })
         } else {
-            res.redirect(`/news/${id}`)
+            return res.status(400).send("Something went wrong")
         }
     }
 })
 
 router.post("/comment/edit", async (req, res) => {
-    const {id, postId, message} = req.body
+    const { id, postId, message } = req.body
 
-    await db.run(`UPDATE comments SET message=? where id=?`, message, id)
-    
-    res.redirect(`/news/${postId}`)
+    if (!Number.isInteger(Number(postId))) {
+        return res.status(400).send("Invalid Post-ID")
+    }
+    if (!Number.isInteger(Number(id))) {
+        return res.status(400).send("Invalid Comment-ID")
+    }
+
+    const comment = await db.get("SELECT * FROM comments WHERE comments.post_id = ? AND comments.id = ?", postId, id)
+
+    if (comment === undefined) {
+        return res.status(400).send("Comment Not Found")
+    } else {
+        if (req.session.userId == comment.author_id) {
+            await db.run(`UPDATE comments SET message=? where id=?`, message, id)
+            res.redirect(`/news/${postId}`)
+        } else {
+            return res.status(400).send("Something went wrong")
+        }
+    }
 })
 
 router.get("/:id/:commentId/delete", async (req, res) => {
@@ -241,9 +240,10 @@ router.get("/:id/:commentId/delete", async (req, res) => {
 
     if (req.session.userId == comment.author_id) {
         await db.run("DELETE FROM comments WHERE comments.id = ?", commentId)
+        res.redirect(`/news/${id}`)
+    } else {
+        return res.status(400).send("Something went wrong")
     }
-
-    res.redirect(`/news/${id}`)
 })
 
 export default router
